Add files via upload
I created this project thanks to the content from the 'Developing Ethical Hacking Tools With Python' from Cybrary. I wrote so far is the functionality that handles command-line arguments. I've encapsulated dysfunctionality into a try and accept statement to catch all possible errors and allow the program to exit in a clean way if it encounters any of these errors.I opened the password file and go line by line and try each word as a password to the ZIP archive.